+++ packages/backend/convex/data_ml/events.ts
@@ -2,8 +2,8 @@ import { latLngToCell } from 'h3-js';
 
 import type { Doc } from '../_generated/dataModel';
 import { query } from '../_generated/server';
+import { __backend_only_guestOrAuthenticatedUser } from '../auth';
 import { eventSeeds } from '../seed';
-import { __backend_only_guestOrAuthenticatedUser } from '../users';
 
 export function latLngToH3Index(lat: number, lng: number, resolution: number = 9): string {
   if (lat < -90 || lat > 90) {
